% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_trees.R
\docType{data}
\name{all_trees}
\alias{all_trees}
\title{Simulated Tree Data with Different Proximity Parameter Value}
\format{
A list of length 12 with each element a sublist containing an
ape \code{multiPhylo} object with 300 rooted trees on 5 tips and the tree categories.
}
\usage{
data(all_trees)
}
\description{
Simulated Tree Data with Different Proximity Parameter Value
}
\keyword{datasets}
