% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apicomplexa.R
\docType{data}
\name{apicomplexa}
\alias{apicomplexa}
\title{Apicomplexa gene trees sample data set.}
\format{
An ape \code{multiPhylo} object with 268 rooted trees on 8 tips.
}
\source{
Chih-Horng Kuo, John P. Wares, Jessica C. Kissinger.
\emph{The Apicomplexan Whole-Genome Phylogeny: An Analysis of Incongruence among Gene Trees
Molecular Biology and Evolution, Volume 25, Issue 12, December 2008, Pages 2689–2698.}
}
\usage{
apicomplexa
}
\description{
Apicomplexa gene trees sample data set.
}
\keyword{datasets}
