% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment_finder.R
\name{assignment_finder}
\alias{assignment_finder}
\title{Efficient Finder of Best Assignment}
\usage{
assignment_finder(P, Q, t = 25)
}
\arguments{
\item{P}{a data matrix of the first category.}

\item{Q}{a data matrix of the second category.}

\item{t}{a numeric value controlling the number of selected best assignments.}
}
\value{
A matrix with each row a unique assignment, starting from the best assignment
and ending with the worst.
}
\description{
Efficient Finder of Best Assignment
}
\keyword{internal}
