% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.cv.tropsvm.R
\name{coef.cv.tropsvm}
\alias{coef.cv.tropsvm}
\title{Extract Optimal Tropical Hyperplane from a cv.tropsvm object}
\usage{
\method{coef}{cv.tropsvm}(object, ...)
}
\arguments{
\item{object}{a fitted \code{"cv.tropsvm"} object.}

\item{\dots}{Not used. Other arguments.}
}
\value{
An output of the apex of the fitted optimal tropical hyperplane.
}
\description{
Obtain the optimal tropical hyperplane in the form of vectors from a cv.tropsvm object.
}
