% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.tropsvm.R
\name{coef.tropsvm}
\alias{coef.tropsvm}
\title{Extract Optimal Tropical Hyperplane from a tropsvm object}
\usage{
\method{coef}{tropsvm}(object, ...)
}
\arguments{
\item{object}{a fitted \code{"tropsvm"} object.}

\item{\dots}{Not used. Other arguments.}
}
\value{
An output of the apex of the fitted optimal tropical hyperplane.
}
\description{
Obtain the optimal tropical hyperplane in the form of vectors from a tropsvm object.
}
