% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dets.R
\name{dets}
\alias{dets}
\alias{trop_det2d}
\alias{trop_det3d}
\alias{trop_det4d}
\alias{trop_det5d}
\title{Tropical Determinants for Small Matrices}
\usage{
trop_det2d(X)

trop_det3d(X)

trop_det4d(X)

trop_det5d(X)
}
\arguments{
\item{X}{a square matrix}
}
\value{
Tropical determinant of a given matrix
}
\description{
Tropical Determinants for Small Matrices
}
\keyword{internal}
