% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linsp_to_poly.R
\name{linsp_to_poly}
\alias{linsp_to_poly}
\title{Convert Tropical Linear Space to Convex Hull}
\usage{
linsp_to_poly(V)
}
\arguments{
\item{V}{a data matrix, of dimension s x e, with each row a basis of tropical linear space.
e is the dimension of the tropical space and s is the dimension of the linear space.}
}
\value{
the return is a e choose s-1 x e matrix.
}
\description{
Convert Tropical Linear Space to Convex Hull
}
\keyword{internal}
