% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.matrix.R
\name{make.matrix}
\alias{make.matrix}
\title{Function to recreate the upper triangular matrix}
\usage{
make.matrix(D, n, tips)
}
\arguments{
\item{D}{the tropical interpretation of a tree as a vector}

\item{n}{a numeric value indicating the the number of leaves of the tree}

\item{tips}{a vector of leaf names}
}
\value{
A cophenetic distance matrix, where each entry is the distance
between any two leaves of the tree
}
\description{
Function to recreate the upper triangular matrix
}
\keyword{internal}
