% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPhylo.to.data.matrix.R
\name{multiPhylo.to.data.matrix}
\alias{multiPhylo.to.data.matrix}
\title{Convert an object in class "multiPhylo" into a data matrix}
\usage{
multiPhylo.to.data.matrix(phy, tipOrder, ncores = 2)
}
\arguments{
\item{phy}{an object in class "multiPhylo"}

\item{tipOrder}{a numeric vector of order of leaf names}

\item{ncores}{a numeric value indicating the number of threads
utilized for multi-cored CPUs. (default: 2)}
}
\value{
a data matrix with the first column of the categories
}
\description{
Convert an object in class "multiPhylo" into a data matrix
}
