% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tropca.R
\name{plot.tropca}
\alias{plot.tropca}
\title{Plot Tropical Principle Component with Projections}
\usage{
\method{plot}{tropca}(x, plab = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{tropca} object.}

\item{plab}{a vector of labels of all points in the given data matrix.
Not needed for unlabeled data. (default: NULL)}

\item{\dots}{Not used. Other arguments to plot}
}
\description{
Visialize second order tropical prinrciple components in \code{tropca}
as tropical triangle with projections of a given data matrix.
}
\examples{
library(Rfast)
n <- 20; e <- 50; s <- 5
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
           rmvnorm(n, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
tropca_fit <- tropca(x)
plot(tropca_fit)
}
\references{
Page, R., Yoshida, R. & Zhang L.
\emph{Tropical principal component analysis on the space of phylogenetic trees.
J. Bioinform., Volume 36, Issue 17, 4590–4598 (2020).}
\url{https://doi.org/10.1093/bioinformatics/btaa564}
}
\author{
Qiwen Kang, Robert Page and Houjie Wang
Maintainer: Houjie Wang \email{wanghoujie6688@gmail.com}
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
