% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tree.to.data.matrix.R
\name{read.tree.to.data.matrix}
\alias{read.tree.to.data.matrix}
\title{Read Newick-formatted trees in two categories into a data matrix}
\usage{
read.tree.to.data.matrix(data.file1, data.file2)
}
\arguments{
\item{data.file1}{A file containing trees in Newick form in a category.}

\item{data.file2}{A file containing trees in Newick form in an assumed different category.}
}
\value{
\code{read.tree.to.data.matrix} has the same return as \code{read.nexus.to.data.matrix}.
}
\description{
Read Newick-formatted trees in two categories into a data matrix
}
