% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_trees.R
\docType{data}
\name{sim_trees}
\alias{sim_trees}
\title{Simulated Tree Data}
\format{
An ape \code{multiPhylo} object with 300 rooted trees on 5 tips.
}
\usage{
data(sim_trees)
}
\description{
Simulated Tree Data
}
\keyword{datasets}
