% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropca.linsp2poly.R
\name{tropca.linsp2poly}
\alias{tropca.linsp2poly}
\title{Tropical Principle Component Analysis by Polytope Converted from Linear Space}
\usage{
tropca.linsp2poly(x, pcs = 2, iteration = list(), ncores = 2)
}
\arguments{
\item{x}{a data matrix, of size n x e, with each row an observation vector.
e is the dimension of the tropical space#'}

\item{pcs}{a numeric value indicating the order of principle component. (default: 2)}

\item{ncores}{a numeric value indicating the number of threads utilized for multi-cored CPUs. (default: 2)}

\item{nsample}{a numeric value indicating the number of samples of MCMC. (default: 1000)}
}
\value{
A list of S3 class \code{"tropca"}, including:
\item{pc}{The principle component as a tropical linear space}
\item{obj}{The tropical PCA objective, the sum of tropical distance from each point to the projection.}
\item{projection}{The projections of all data points.}
\item{type}{The geometry of principle component.}
}
\description{
Approximates the principle component as a tropical polytope converted from tropical linear space for a given data matrix
via MCMC and return the results as an object of class \code{tropca}.
}
\examples{
\dontrun{
library(Rfast)
n <- 50; e <- 50; s <- 5
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
           rmvnorm(n, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
tropca_fit <- tropca.linsp2poly(x)
plot(tropca_fit)
}

}
\references{
Page, R., Yoshida, R. and Zhang L. (2020)
\emph{Tropical Principal Component Analysis on the Space of Phylogenetic Trees,
 Bioinformatics, Volume 36, Issue 17, 4590–4598.}
\url{https://doi.org/10.1093/bioinformatics/btaa564}.
}
\author{
Houjie Wang
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
