% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropca.obj.R
\name{tropca.obj}
\alias{tropca.obj}
\title{Compute Tropical PCA Objective}
\usage{
tropca.obj(pc, x_list, cl)
}
\arguments{
\item{pc}{a matrix of principle components}

\item{x_list}{a list of vectors}

\item{cl}{cluster for parallel computing}
}
\description{
Compute Tropical PCA Objective
}
