% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropca.R
\name{tropca}
\alias{tropca}
\title{Tropical Principle Components Analysis}
\usage{
tropca(x, pcs = 3, nsample = 1000, ncores = 2)
}
\arguments{
\item{x}{a data matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{pcs}{a numeric value indicating the number of principle components. (default: 3)}

\item{nsample}{a numeric value indicating the number of samples of MCMC. (default: 1000)}

\item{ncores}{a numeric value indicating the number of threads utilized for multi-cored CPUs. (default: 2)}
}
\value{
A list of S3 class \code{"tropca"} containing the fitted model, including:
\item{pc}{The approximated principle components as a tropical polytope.}
\item{obj}{The tropical PCA objective, the sum of tropical distances from points to their projections on the principle components.}
\item{projection}{The projections of points.}
\item{type}{The geometry of principle components.}
}
\description{
Approximate the principle components as a polytope for a given data matrix in the settings of
tropical geometry via MCMC and return the results as an object of class tropca.
}
\examples{
\dontrun{
library(Rfast)
n <- 50; e <- 50; s <- 5
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
           rmvnorm(n, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
tropca_fit <- tropca(x)
plot(tropca_fit)
}

}
\author{
Houjie Wang
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
