% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropdet.R
\name{tropdet}
\alias{tropdet}
\title{Tropical Determinant of a Matrix}
\usage{
tropdet(x)
}
\arguments{
\item{x}{a square matrix}
}
\value{
The determinant of the given matrix,
}
\description{
Compute the tropical determinant for a given matrix. This is equivalent to
solving an assignment problem.
}
\examples{
R <- matrix(sample(1:9, 9), nrow = 3)
tropdet(R)
}
