% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troproj.R
\name{troproj}
\alias{troproj}
\title{Projection on Tropical Polytope}
\usage{
troproj(x, tconv)
}
\arguments{
\item{x}{a data vector.}

\item{tconv}{a data matrix, of dimension e x s, with each column a vertex of tropical polytope.
e is the dimension of the tropical space and s is the dimension of the polytope.}
}
\value{
A projected vector on the given tropical polytope.
}
\description{
\code{troproj} computes a vector projection on a given tropical polytope.
}
\examples{

# Generate a tropical polytope consisting of three trees each with 5 leaves
library(ape)
pltp <- sapply(1: 3, function(i){vec.fun(rcoal(5))})
# Generate an observation and vectorize it
tree <- rcoal(5)
tree_vec <- vec.fun(tree)
troproj(tree_vec, pltp)

}
\author{
Robert Page and Houjie Wang
}
