% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropsvm.R
\name{tropsvm}
\alias{tropsvm}
\title{Tropical Support Vector Machines}
\usage{
tropsvm(
  x,
  y,
  auto.assignment = FALSE,
  accuracy = FALSE,
  assignment = NULL,
  ind = NULL,
  newx = NULL,
  newy = NULL
)
}
\arguments{
\item{x}{a data matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{y}{a response vector with one label for each row/component of x.}

\item{auto.assignment}{a logical value indicating if to provide an \code{assignment} by user.
If \code{FALSE}, an input is needed from the user, otherwise the function automatically
finds a good assignment.(default: FALSE)}

\item{accuracy}{a logical value indicating if return accuracy on test data and test label
\code{newx} and \code{newy} only. Note that if \code{TRUE}, the test data and test label
\code{newx} and \code{newy} should be provided. Users are more recommended to use its
default value as testing by prediction can be done via \code{predict.tropsvm}
more formally. (default: FALSE)}

\item{assignment}{a numeric vector indicating the sectors of tropical hyperplane that the
data will be assigned to. The first and third elements in the \code{assignment} are the coordinates of
an observed point in data matrix \code{x} believed from the first category where the maximum and second maximum
of the vector addition between the fitted optimal tropical hyperplane and the point itelf are achieved.
The meanings for the second and the fourth element in the \code{assignment} are the same
but for the points in the second category. Namely, the first and second values in the \code{assignment}
are the indices of sectors where the two point cloud will be assigned. Not needed when \code{auto.assignment = TRUE}. (default: NULL)}

\item{ind}{a numeric value or a numeric vector ranging from 1 to 70 indicating which classification method
will be used. There are 70 different classification methods. The different classification methods are proposed to resolve
the issue when points fall on the intersections of sectors. Users can have personal choices if better knowledge is assumed. (default: 1)}

\item{newx}{the same as "x" but only needed in \code{cv.tropsvm}, which is used as validation data. (default: \code{NULL})}

\item{newy}{the same as "y" but only needed in \code{cv.tropsvm}, which is used as validation labels (default: \code{NULL})}
}
\value{
An object with S3 class \code{"tropsvm"}.
\item{coef}{The vector of the fitted optimal tropical hyperplane}
\item{assignment}{The user-input \code{assignment}}
\item{method index}{The user-input \code{ind}}
\item{levels}{The name of each category, consistent with categories in \code{y}}
}
\description{
Fit a discriminative two-class classifier via linear programming defined by the tropical
hyperplane which maximizes the minimum tropical distance from data points
to itself in order to separate the data points into sectors (half-spaces)
in the tropical projective torus.
}
\examples{

# data generation
library(Rfast)
e <- 100; n = 100; N = 100; s = 10
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
          rmvnorm(n, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
y <- as.factor(c(rep(1, n), rep(2, n)))
newx <- rbind(rmvnorm(N, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
             rmvnorm(N, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
newy <- as.factor(rep(c(1, 2), each = N))

# train the tropical svm
tropsvm_fit <- tropsvm(x, y, auto.assignment = TRUE, ind = 1)

coef(tropsvm_fit)

# test with new data
pred <- predict(tropsvm_fit, newx)

# check with accuracy
table(pred, newy)

# compute testing accuracy
sum(pred == newy)/length(newy)

}
\references{
Tang, X., Wang, H. and Yoshida, R. (2020)
\emph{Tropical Support Vector Machine and its Applications to Phylogenomics}
\url{https://arxiv.org/pdf/2003.00677.pdf}
}
\seealso{
\code{predict}, \code{coef} and the \code{cv.tropsvm} function.
}
\author{
Houjie Wang and Kaizhang Wang
Maintainer: Houjie Wang \email{whj666@uw.edu}
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
