% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_fun.R
\name{vec_fun}
\alias{vec_fun}
\title{Vectorize a distance matrix}
\usage{
vec_fun(x)
}
\arguments{
\item{x}{A distance matrix between each pair of leaves.}
}
\description{
Vectorize a distance matrix
}
\examples{
vec_fun(m)

}
