% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.tropsvm.R
\name{cv.tropsvm}
\alias{cv.tropsvm}
\title{Cross-Validation for Tropical Support Vector Machines}
\usage{
cv.tropsvm(x, y, parallel = TRUE, nfold = 10, nassignment = 10, ncores = 2)
}
\arguments{
\item{x}{a data matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{y}{a response vector with one label for each row/component of x.}

\item{parallel}{a logical value indicating if parallel computing should be used. (default: TRUE)}

\item{nfold}{a numeric value of the number of data folds for cross-validation. (default: 10)}

\item{nassignment}{a numeric value indicating the size of the parameter grid of assignments. (default: 10)}

\item{ncores}{a numeric value indicating the number of threads utilized for multi-cored CPUs. (default: 2)}
}
\value{
An object with S3 class \code{"cv.tropsvm"}.
\item{coef}{The vector of the fitted optimal tropical hyperplane.}
\item{assignment}{The best \code{assignment} tuned by cross-validation.}
\item{method index}{The best classification method indices \code{ind} tuned by cross-validation.}
\item{levels}{The name of each category, consistent with categories in \code{y}.}
\item{accuracy}{The validation accuracy for each validation set.}
\item{nfold}{The number of folds used in cross-validation}
}
\description{
Conduct k-fold cross validation for tropsvm and return an object \code{"cv.tropsvm"}.
}
\examples{

# data generation
library(Rfast)
e <- 100; n = 100; N = 100; s = 10
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
          rmvnorm(n, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
y <- as.factor(c(rep(1, n), rep(2, n)))
newx <- rbind(rmvnorm(N, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
             rmvnorm(N, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
newy <- as.factor(rep(c(1, 2), each = N))

# train the tropical svm
cv_tropsvm_fit <- cv.tropsvm(x, y)

summary(cv_tropsvm_fit)
coef(cv_tropsvm_fit)

# test with new data
pred <- predict(cv_tropsvm_fit, newx)

# check with accuracy
table(pred, newy)

# compute testing accuracy
sum(pred == newy)/length(newy)

}
\references{
Tang, X., Wang, H. and Yoshida, R. (2020)
\emph{Tropical Support Vector Machine and its Applications to Phylogenomics}
\url{https://arxiv.org/pdf/2003.00677.pdf}
}
\seealso{
\code{summary}, \code{predict}, \code{coef} and the \code{tropsvm} function.
}
\author{
Houjie Wang and Kaizhang Wang
Maintainer: Houjie Wang \email{whj666@uw.edu}
}
