% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPhylo.to.data.matrix.R
\name{multiPhylo.to.data.matrix}
\alias{multiPhylo.to.data.matrix}
\title{Convert an object in class "multiPhylo" into a data matrix}
\usage{
multiPhylo.to.data.matrix(phy, tipOrder)
}
\arguments{
\item{phy}{An object in class "multiPhylo"}

\item{tipOrder}{Order of leaf names}
}
\value{
A data matrix with the first column of the categories
}
\description{
Convert an object in class "multiPhylo" into a data matrix
}
