% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiPhylo.to.data.matrix.R
\name{multiPhylo.to.data.matrix}
\alias{multiPhylo.to.data.matrix}
\title{Convert an object in class "multiPhylo" into a data matrix}
\usage{
multiPhylo.to.data.matrix(trees, tipOrder)
}
\arguments{
\item{trees}{An object in class "multiPhylo" containing trees}

\item{tipOrder}{Order of leaf names}
}
\description{
Convert an object in class "multiPhylo" into a data matrix
}
\examples{
multiPhylo.to.data.matrix(trees, tipOrder)

}
