% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tropca.R
\name{plot.tropca}
\alias{plot.tropca}
\title{Plot the Tropical Principal Components with Data Projections}
\usage{
\method{plot}{tropca}(x, plab = NULL, ...)
}
\arguments{
\item{x}{a fitted \code{tropca} object.}

\item{plab}{a vector of labels of all points in the given data matrix.
Not needed for unlabeled data. (default: NULL)}

\item{\dots}{Not used. Other arguments to plot}
}
\description{
Visialize the second order tropical prinrciple components in \code{tropca}
as a tropical triangle with projections on a two-dimensional plot via tropical isometry.
}
\author{
Robert Page and Houjie Wang
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
