% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tropca.R
\name{plot.troppca}
\alias{plot.troppca}
\title{Plot the Tropical Principal Components with Data Projections}
\usage{
\method{plot}{troppca}(x, plab = NULL, fw = FALSE, ...)
}
\arguments{
\item{x}{a fitted \code{troppca} object.}

\item{plab}{a vector of labels of all points in the given data matrix.
Not needed for unlabeled data. (default: NULL)}

\item{fw}{a logical variable to determine if to add
Fermat-Weber point of the data projection. (default: FALSE)}

\item{\dots}{Not used. Other arguments to plot}
}
\value{
\code{plot.troppca} does not return anything other than the plot.
}
\description{
Visualize the second order tropical principle components in \code{troppca}
as a tropical triangle with projections on a two-dimensional plot via tropical isometry.
}
