% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytope_iso.R
\name{polytope_iso}
\alias{polytope_iso}
\title{Dimension Reduction by Isometry}
\usage{
polytope_iso(D, P)
}
\arguments{
\item{D}{a data matrix of size s x e whose rows are vertices of a tropical polytope.}

\item{P}{a vector of a point in row spaces of \code{D}.}
}
\value{
A vector of a point corresponding to \code{P} in column span of \code{D}.
}
\description{
Dimension Reduction by Isometry
}
\keyword{Map}
\keyword{a}
\keyword{column}
\keyword{in}
\keyword{internal}
\keyword{its}
\keyword{of}
\keyword{points}
\keyword{polytope}
\keyword{row}
\keyword{space}
\keyword{space.}
\keyword{to}
\keyword{tropical}
