% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.nexus.to.data.matrix.R
\name{read.nexus.to.data.matrix}
\alias{read.nexus.to.data.matrix}
\title{Read trees in two categories into a data matrix}
\usage{
read.nexus.to.data.matrix(data.file1, data.file2)
}
\arguments{
\item{data.file1}{A data set with trees from one category.}

\item{data.file2}{A data set with trees from the other category.}
}
\description{
Read trees in two categories into a data matrix
}
\examples{
read.nexus.to.data.matrix(data.file1, data.file2)

}
