% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.tree.to.data.matrix.R
\name{read.tree.to.data.matrix}
\alias{read.tree.to.data.matrix}
\title{Read trees in two categories into a data matrix}
\usage{
read.tree.to.data.matrix(data.file1, data.file2)
}
\arguments{
\item{data.file1}{A set of trees in the format of Newick strings from one category.}

\item{data.file2}{a set of trees in the format of Newick strings from another category.}
}
\description{
Read trees in two categories into a data matrix
}
\examples{
read.tree.to.data.matrix(data.file1, data.file2)

}
