% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropca.linsp.R
\name{tropca.linsp}
\alias{tropca.linsp}
\title{Tropical Principle Components Analysis by Tropical Linear Space}
\usage{
tropca.linsp(x, pcs = 3, size = 100, ncores = 2, exhaust = FALSE)
}
\arguments{
\item{x}{a data matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{pcs}{a numeric value indicating the number of principle components. (default: 3)}

\item{size}{a numeric value indicating the number of tropical linear spaces for where the principle components are searched. (default: 100)}

\item{ncores}{a numeric value indicating the number of threads utilized for multi-cored CPUs. (default: 2)}

\item{exhaust}{a logical variable indicating if to search for all possible combinations of the data matrix.(default: FALSE)}
}
\value{
A list of S3 class \code{"tropca"} containing the fitted model, including:
\item{pc}{The approximated principle components as a tropical linear space}
\item{obj}{The tropical PCA objective, the sum of tropical distances from points to their projections on the principle components.}
\item{projection}{The projections of points.}
\item{type}{The geometry of principle components.}
}
\description{
Approximate the principle components as a tropical linear space for a given data matrix in the settings of
tropical geometry and return the results as an object of class tropca.
}
\examples{
\dontrun{
library(Rfast)
n <- 100; e <- 10; sig2 <- 1; pcs <- 3
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(sig2, e)))
tropca_fit <- tropca.linsp(x)
}

}
\author{
Houjie Wang
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
