% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropca.R
\name{tropca}
\alias{tropca}
\title{Tropical Principle Components Analysis}
\usage{
tropca(x, pcs = 3, rep = 2)
}
\arguments{
\item{x}{a data matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{pcs}{a numeric value indicating the number of principle components. (default: 3)}

\item{rep}{a numeric value indicating the number of repetitions of MCMC. (default: 2)}
}
\value{
A list of S3 class \code{"tropca"} containing the fitted model, including:
\item{pc}{The principle components.}
\item{obj}{The sum of tropical distances from each point to its projection on the principle components.}
\item{projection}{The projections of given data set.}
}
\description{
Approximate the principle components as a polytope for a given data matrix in the settings of
tropical geometry via MCMC and return the results as an object of class tropca.
}
\examples{
library(Rfast)
n <- 20; e <- 50; s <- 5
x <- rbind(rmvnorm(n, mu = c(5, -5, rep(0, e-2)), sigma = diag(s, e)),
           rmvnorm(n, mu = c(-5, 5, rep(0, e-2)), sigma = diag(s, e)))
tropca_fit <- tropca(x)
plot(tropca_fit)
}
\references{
Page, R., Yoshida, R. & Zhang L.
\emph{Tropical principal component analysis on the space of phylogenetic trees.
J. Bioinform., Volume 36, Issue 17, 4590–4598 (2020).}
\url{https://doi.org/10.1093/bioinformatics/btaa564}
}
\author{
Houjie Wang

Maintainer: Houjie Wang \email{wanghoujie6688@gmail.com}
}
\keyword{Data}
\keyword{Geometry,}
\keyword{Learning,}
\keyword{Non-Euclidean}
\keyword{Supervised}
\keyword{Tropical}
