% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tropFW.R
\name{tropFW}
\alias{tropFW}
\title{Tropical Fermat-Weber Point}
\usage{
tropFW(x)
}
\arguments{
\item{x}{a data matrix, of dimension nobs x nvars; each row is an observation vector.}
}
\value{
A list containing:
\item{fw}{The fermat-weber point.}
\item{distsum}{The sum of distance from each observation to the fermat-weber point.}
}
\description{
Compute the tropical fermat-weber point for a given data set such that the sum of tropical distance to each point is minimized.
}
\examples{
x = matrix(rnorm(100), ncol = 10)
tropFW(x)

}
\references{
Lin, B.,Sturmfels, B., Tang, X. and Yoshida, R. (2015)
\emph{Convexity in Tree Spaces, SIAM Journal on Discrete Mathematics, Vol. 31(3) 2015–2038}
}
\author{
Houjie Wang

Maintainer: Houjie Wang \email{wanghoujie6688@gmail.com}
}
\keyword{Fermat-Weber}
\keyword{Geometry,}
\keyword{Point}
\keyword{Tropical}
