% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troproj.linsp.R
\name{troproj.linsp}
\alias{troproj.linsp}
\title{Projection on Tropical Linear Space}
\usage{
troproj.linsp(x, V)
}
\arguments{
\item{x}{a data vector or a data matrix with each row an observation.}

\item{V}{a data matrix, of dimension s x e, with each row a basis of tropical linear space.
e is the dimension of the tropical space and s is the dimension of the linear space.}
}
\description{
troproj.linsp computes a vector projection on a given tropical linear space
}
