% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troproj.R
\name{troproj}
\alias{troproj}
\title{Projection on Tropical Polytope}
\usage{
troproj(x, tconv)
}
\arguments{
\item{x}{x a data matrix, of dimension nobs x nvars; each row is an observation vector.}

\item{tconv}{a data matrix, of dimension nvars x s; each column is a basis of tropical polytope;
s is the dimension of the polytope.}
}
\value{
The projected vector on the given tropical polytope.
}
\description{
\code{troproj} computes a projection onto a given tropical polytope.
}
\examples{

# Generate a tropical polytope consisting of three trees each with 5 leaves
library(ape)
pltp <- sapply(1: 3, function(i){vec.fun(rcoal(5))})
# Generate an observation and vectorize it
tree <- rcoal(5)
tree_vec <- vec.fun(tree)
troproj(tree_vec, pltp)

}
\references{
Page, R., Yoshida, R. & Zhang L.
\emph{Tropical principal component analysis on the space of phylogenetic trees.
J. Bioinform., Volume 36, Issue 17, 4590–4598 (2020).}
\url{https://doi.org/10.1093/bioinformatics/btaa564}
}
\author{
Qiwen Kang and Houjie Wang

Maintainer: Houjie Wang \email{wanghoujie6688@gmail.com}
}
